feat: add Next.js configuration with MDX support
- Introduced a new configuration file for Next.js to support MDX file extensions.
- Integrated remark-gfm for GitHub Flavored Markdown and rehype-prism-plus for syntax highlighting.
- Configured image remote patterns and page extensions for better compatibility.
